Added Basic Acceptance Tests
Needed to figure out a way for checking whether a query was resolved.
Here are some ideas.

Make  query and check that all fields have resolved not just to valid
values, but to be valid, non-empty values. This has to stop somewhere.
What about schemas where not every value is non-empty?

Another idea would be to snapshot the shape of the data resulting from
the query, including array entries.

The only variant data in the snapshot is the cursor. An easy way would
be to load the same mock data into core and make queries against both
resolvers. A separate test for pagination will be needed. This is the
approach I've chosen.