Fixed Missing list and kind in getTracks
The PlaybackAlbumArtwork gets a list of tracks which are consumed by a
PlaybackArtwork. The playback artwork enhances the list of tracks with
list and kind information based on its props. This getTracks
implementation assumed that it would only be used in a context where it
was enhanced.

In the Album page, it is used without enhancement. This issue could have
been avoided by using a better <PlaybackArtwork> abstraction or more
tightly binding the types which would have been easier to do with a
simple QueueItem type.